requirements.txt
fastapi
uvicorn
python-multipart
pillow
# Get CPU-only torch to keep image small
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu